+++ src/main/java/com/dozenflow/be/label/LabelRepository.java
@@ -1,0 +1,6 @@
+package com.dozenflow.be.label;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface LabelRepository extends JpaRepository<Label, Long> {
+}