+++ src/main/java/com/dozenflow/be/task/mapper/TaskMapper.java
@@ -1,12 +1,23 @@
 package com.dozenflow.be.task.mapper;
 
+import com.dozenflow.be.label.dto.LabelResponseDTO;
+import com.dozenflow.be.label.mapper.LabelMapper;
 import com.dozenflow.be.task.Task;
 import com.dozenflow.be.task.dto.TaskRequestDTO;
 import com.dozenflow.be.task.dto.TaskResponseDTO;
 import org.springframework.stereotype.Component;
 
+import java.util.Comparator;
+import java.util.List;
+
 @Component
 public class TaskMapper {
+
+    private final LabelMapper labelMapper;
+
+    public TaskMapper(LabelMapper labelMapper) {
+        this.labelMapper = labelMapper;
+    }
 
     public Task toEntity(TaskRequestDTO dto) {
         Task task = new Task();
@@ -19,9 +30,14 @@ public class TaskMapper {
     }
 
     public TaskResponseDTO toResponseDTO(Task entity) {
+        List<LabelResponseDTO> labels = entity.getLabels().stream()
+                .map(labelMapper::toResponseDTO)
+                .sorted(Comparator.comparing(LabelResponseDTO::id))
+                .toList();
+
         return new TaskResponseDTO(
                 entity.getId(), entity.getTitle(), entity.getDescription(),
-                entity.getStatus(), entity.getTaskOrder(), entity.getDueDate()
+                entity.getStatus(), entity.getTaskOrder(), entity.getDueDate(), labels
         );
     }
 }